backend/requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
httpx
langgraph
langchain-google-genai
langchain-community
ddgs